moved loading and interceptor to assets
